requirements.txt
asgiref==3.8.1
black==24.4.0
click==8.1.7
colorama==0.4.6
Django==4.2.14
django-debug-toolbar==4.4.6
django-model-utils==4.5.0
djangorestframework==3.15.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
